+++ recipe/thirdpartypasswordless/smsdelivery/twilioService/main.go
@@ -22,18 +22,18 @@ import (
 	"github.com/supertokens/supertokens-golang/supertokens"
 )
 
-func MakeTwilioService(config smsdelivery.TwilioTypeInput) (smsdelivery.SmsDeliveryInterface, error) {
+func MakeTwilioService(config smsdelivery.TwilioServiceConfig) (*smsdelivery.SmsDeliveryInterface, error) {
 	plessServiceImpl, err := twilioService.MakeTwilioService(config)
 
 	if err != nil {
-		return smsdelivery.SmsDeliveryInterface{}, err
+		return nil, err
 	}
 
 	sendSms := func(input smsdelivery.SmsType, userContext supertokens.UserContext) error {
 		return (*plessServiceImpl.SendSms)(input, userContext)
 	}
 
-	return smsdelivery.SmsDeliveryInterface{
+	return &smsdelivery.SmsDeliveryInterface{
 		SendSms: &sendSms,
 	}, nil
 }